refactor: 데이터 관련 클래스를 model로 묶어주기
