+++ src/BridgeGameControl.js
@@ -1,9 +1,7 @@
 /* eslint-disable class-methods-use-this */
 const { Console } = require('@woowacourse/mission-utils');
 const InputView = require('./InputView');
-const BridgeSize = require('./BridgeSize');
-const Moving = require('./Moving');
-const GameCommand = require('./GameCommand');
+const { BridgeSize, Moving, GameCommand } = require('./model');
 /**
  * 다리 건너기 게임을 관리하는 클래스
  */
