+++ src/model/index.js
@@ -1,0 +1,5 @@
+const BridgeSize = require('./BridgeSize');
+const Moving = require('./Moving');
+const GameCommand = require('./GameCommand');
+
+module.exports = { BridgeSize, Moving, GameCommand };
